Added an If statement
The if statement allows nondim to run even if T is zero, as it sets FM to zero in that case. CT will still give you the correct values, and FM will give correct values for T > 0, when T < 0 it normally does the square root of a negative (for FM), but that is fixed so that it now just sets FM to 0 in that case.